+++ node/libs/roles/src/node/messages.rs
@@ -1,7 +1,10 @@
-use zksync_consensus_crypto::{keccak256, ByteFmt, Text, TextFmt};
-use zksync_consensus_utils::enum_util::{BadVariantError, Variant};
+use anyhow::Context as _;
 
-use crate::node;
+use zksync_consensus_crypto::{keccak256::Keccak256, ByteFmt, Text, TextFmt};
+use zksync_consensus_utils::enum_util::{BadVariantError, Variant};
+use zksync_protobuf::{read_required, required, ProtoFmt};
+
+use crate::{node, proto::node as proto};
 
 /// The ID for an authentication session.
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -18,7 +21,7 @@ pub enum Msg {
 impl Msg {
     /// Get the hash of this message.
     pub fn hash(&self) -> MsgHash {
-        MsgHash(keccak256::Keccak256::new(&zksync_protobuf::canonical(self)))
+        MsgHash(Keccak256::new(&zksync_protobuf::canonical(self)))
     }
 }
 
@@ -29,6 +32,23 @@ impl Variant<Msg> for SessionId {
     fn extract(msg: Msg) -> Result<Self, BadVariantError> {
         let Msg::SessionId(this) = msg;
         Ok(this)
+    }
+}
+
+impl ProtoFmt for Msg {
+    type Proto = proto::Msg;
+    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
+        use proto::msg::T;
+        Ok(match required(&r.t)? {
+            T::SessionId(r) => Self::SessionId(SessionId(r.clone())),
+        })
+    }
+    fn build(&self) -> Self::Proto {
+        use proto::msg::T;
+        let t = match self {
+            Self::SessionId(x) => T::SessionId(x.0.clone()),
+        };
+        Self::Proto { t: Some(t) }
     }
 }
 
@@ -52,8 +72,26 @@ impl<V: Variant<Msg> + Clone> Signed<V> {
     }
 }
 
+impl<V: Variant<Msg> + Clone> ProtoFmt for Signed<V> {
+    type Proto = proto::Signed;
+    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
+        Ok(Self {
+            msg: V::extract(read_required::<Msg>(&r.msg).context("msg")?)?,
+            key: read_required(&r.key).context("key")?,
+            sig: read_required(&r.sig).context("sig")?,
+        })
+    }
+    fn build(&self) -> Self::Proto {
+        Self::Proto {
+            msg: Some(self.msg.clone().insert().build()),
+            key: Some(self.key.build()),
+            sig: Some(self.sig.build()),
+        }
+    }
+}
+
 /// The hash of a message.
-pub struct MsgHash(pub(super) keccak256::Keccak256);
+pub struct MsgHash(pub(crate) Keccak256);
 
 impl ByteFmt for MsgHash {
     fn encode(&self) -> Vec<u8> {
