+++ node/libs/roles/src/node/mod.rs
@@ -1,5 +1,4 @@
 //! Node role implementation.
-mod conv;
 mod keys;
 mod messages;
 mod testonly;
